Fix/improve pywhispercpp (#87)
* Use samples_to_seconds to document the operation

* Provide spaces before words, since whisper.cpp strips them

* route whispercpp logs to sys.stderr

* Use the callback interface to receive words

* Use more cores when available

* use seconds_to_samples

* ignore empty words

* Apply window offset to word timestamp

* Further tune whisper.cpp with language, prompt, and reset context each time

+++ verbatim/audio/audio.py
@@ -72,7 +72,7 @@ def samples_to_seconds(index: int) -> float:
 
 
 def seconds_to_samples(seconds: float) -> int:
-    return seconds * 16000
+    return int(seconds * 16000)
 
 
 def timestr_to_samples(timestr: str, sample_rate: int = 16000) -> int:
